requirements.txt
requirements.txt
streamlit
google-generativeai
gnews
pandas
openpyxl
python-dotenv

